refactor(web): split keys-section giant file (S8, #1035)
keys-section.tsx 1328 → 416 lines: the create/edit sheet form (model
policy editor + site scope picker + RHF body) moves to
key-sheet-form.tsx, the table cell renderers to key-cells.tsx, and the
shared API types / query keys / zod schemas / value mappers to
key-form-shared.ts. keys-section.tsx re-exports the moved components so
the four focused test files keep importing from the section barrel —
41/41 domain tests pass unchanged. Zero behavior change; verified on a
seeded instance with desktop+mobile screenshots of the create sheet.
